MicrosoftEdgeReadingSectionV2.js
An update to the Microsoft Edge browser sniffing script, adding support for Microsoft Edge version 83. This is still just a template, and does not work at all yet. I also added template sections for the non-existent 0.x to 19.x releases to clarify that they don't exist, and that releases started with version 20.